Adding env variables and seperation of Pbft apis
