+++ package.json
@@ -15,6 +15,7 @@
     "axios": "^1.6.2",
     "classnames": "^2.3.2",
     "d3": "^7.8.5",
+    "dotenv": "^16.4.5",
     "lodash": "^4.17.21",
     "object-hash": "^3.0.0",
     "re-resizable": "^6.9.11",
